chore(entities): refacto entities folder to reduce circular dependancies
Changelog
* Refacto entities folder
