+++ navitia_client/entities/base_entity.py
