+++ navitia_client/entities/equipment_reports.py
@@ -1,0 +1,10 @@
+from dataclasses import dataclass
+
+from .line import Line
+from .equipment import StopAreaEquipments
+
+
+@dataclass
+class EquipmentReports:
+    line: Line
+    stop_area_equipments: StopAreaEquipments
