+++ navitia_client/entities/network.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+from .base_entity import BaseEntity
+
+
+@dataclass
+class Network(BaseEntity):
+    pass
