+++ navitia_client/entities/note.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Note:
+    id: str
+    value: str
