+++ navitia_client/entities/poi.py
@@ -1,0 +1,16 @@
+from dataclasses import dataclass
+
+from .base_entity import BaseEntity
+from .stand import Stands
+
+
+@dataclass
+class POIType(BaseEntity):
+    pass
+
+
+@dataclass
+class POI(BaseEntity):
+    label: str
+    poi_type: POIType
+    stands: Stands
